+++ src/webserver/database/sqlpage_functions/functions/client_ip.rs
@@ -1,0 +1,5 @@
+use super::*;
+
+pub(super) async fn client_ip(request: &RequestInfo) -> Option<String> {
+    Some(request.client_ip?.to_string())
+}
