+++ src/webserver/database/sqlpage_functions/functions/cookie.rs
@@ -1,0 +1,5 @@
+use super::*;
+
+pub(super) async fn cookie<'a>(request: &'a RequestInfo, name: Cow<'a, str>) -> Option<Cow<'a, str>> {
+    request.cookies.get(&*name).map(SingleOrVec::as_json_str)
+}
