+++ src/webserver/database/sqlpage_functions/functions/headers.rs
@@ -1,0 +1,5 @@
+use super::*;
+
+pub(super) async fn headers(request: &RequestInfo) -> String {
+    serde_json::to_string(&request.headers).unwrap_or_default()
+}
