+++ src/webserver/database/sqlpage_functions/functions/path.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+/// Returns the path component of the URL of the current request.
+pub(super) async fn path(request: &RequestInfo) -> &str {
+    &request.path
+}
