+++ src/webserver/database/sqlpage_functions/functions/protocol.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+/// Returns the protocol of the current request (http or https).
+pub(super) async fn protocol(request: &RequestInfo) -> &str {
+    &request.protocol
+}
