+++ src/webserver/database/sqlpage_functions/functions/request_method.rs
@@ -1,0 +1,5 @@
+use super::*;
+
+pub(super) async fn request_method(request: &RequestInfo) -> String {
+    request.method.to_string()
+}
