+++ src/webserver/database/sqlpage_functions/functions/uploaded_file_name.rs
@@ -1,0 +1,13 @@
+use super::*;
+
+pub(super) async fn uploaded_file_name<'a>(
+    request: &'a RequestInfo,
+    upload_name: Cow<'a, str>,
+) -> Option<Cow<'a, str>> {
+    let fname = request
+        .uploaded_files
+        .get(&*upload_name)?
+        .file_name
+        .as_ref()?;
+    Some(Cow::Borrowed(fname.as_str()))
+}
