+++ src/webserver/database/sqlpage_functions/functions/version.rs
@@ -1,0 +1,5 @@
+
+/// Returns the version of the sqlpage that is running.
+pub(super) async fn version() -> &'static str {
+    env!("CARGO_PKG_VERSION")
+}
