+++ src/webserver/database/sqlpage_functions/mod.rs
@@ -1,4 +1,3 @@
-mod function_definition_macro;
 mod function_traits;
 pub(super) mod functions;
 mod http_fetch_request;
